added {transports: ['websocket']}
Added {transports: ['websocket']} to the streamlabs const (thanks to Ioangogo)
Changed it to try fix the sudden loss of response from websockets, but didn't change anything (it fixed itself later (probably something Streamlabs did on their end)). Apparently it's what it was supposed to be this whole time according to the documentation, so I'm keeping it in.